% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{add_zero_edges}
\alias{add_zero_edges}
\title{Add Zero Edges}
\usage{
add_zero_edges(tree, node = 1:Nnode(tree))
}
\arguments{
\item{tree}{an object of class 'phylo'.}

\item{node}{integer object specifying nodes to add zero-length edges to, where 1 = root.}
}
\value{
An object of class phylo.
}
\description{
Function to add zero length edges to nodes of a phylo object.
}
\details{
This function adds zero-length edges to a user-specified set of internal nodes of a phylo object (defualt = all internal nodes). This is useful if you need to constrain internal nodes to a particular state during ancestral state estimation.
}
\examples{
t1 <- rtree(10)
par(mfrow = c(2, 1))
plot(t1, no.margin = T)
t2 <- add_zero_edges(t1)
plot(t2, no.margin = T)
par(mfrow = c(1, 1))

}
