% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{anc_timeslice}
\alias{anc_timeslice}
\title{Anc timeslice}
\usage{
anc_timeslice(tree, x, anc = NULL, slices)
}
\arguments{
\item{tree}{an object of class 'phylo'.}

\item{x}{vector of tip values for species; names(x) should be the species names.}

\item{anc}{vector of internal node values. If not specified, these will be estimated using BM.}

\item{slices}{integer specifying the number of timeslices to use.}
}
\value{
Returns a list of length slices + 1, containing the trait value at the root and for each edge intersecting each timeslice.
}
\description{
Function that estimates ancestral trait values along each edge of a phylogenetic tree, based on a continuous trait and a user-specified number of time slices.
}
\details{
Modified from phytools blog: see https://blog.phytools.org/2017/01/extracting-reconstructed-trait-values.html. If node values are suplied using the 'anc' argument, internal node values are fixed using add_zero_edges(). The function then uses phytools::fastAnc to interpolate between fixed node values.
}
\examples{
t1 <- pbtree(n = 10)
## Simulate trait
x <- fastBM(t1)
anc_timeslice(t1, x, slices = 10)
}
