% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{Brier}
\alias{Brier}
\title{Brier Score}
\usage{
Brier(prediction, truth)
}
\arguments{
\item{prediction}{a numeric vector of length equal to the number of possible outcomes. Each element represents the probability (between 0 and 1) of a possible outcome.}

\item{truth}{A vector of actual outcomes, with 0 indicating the outcome did not occur and 1 indicating it did.}
}
\value{
A Brier score between 0 and 1.
}
\description{
Function to calculate the Brier score given a vector of predicted probabilities and a vector of actual outcomes.
}
\details{
to do!
}
\examples{
## Example marginal ancestral state estimate for a node
x <- runif(n = 4, min = 0, max = 10)
x <- x/sum(x)
names(x) <- 1:4
x
## Example true data
y <- sample(c(1, 0, 0, 0))
names(y) <- 1:4
y
## Calculate Brier score
Brier(prediction = x, truth = y)

}
