% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{clade_support}
\alias{clade_support}
\title{Calculate support values for a specific tree given a sample of trees}
\usage{
clade_support(trees, con, digits = 2)
}
\arguments{
\item{trees}{a tree object of class 'multiPhylo'.}

\item{con}{a tree object of class 'phylo'.}

\item{digits}{integer indicating the number of decimal places to round to.}
}
\value{
A tree object of class 'phylo' with node labels equivalent to clade_support.
}
\description{
Function to calculate support values for a specific tree (e.g. a consensus tree, maximum clade credibility tree, maximum likelihood tree etc.) given a sample of trees (e.g. posterior trees, bootstrap replicate trees, most parsimonious trees).
}
\details{
This function is a wrapper for the ape function prop.clades.
}
\examples{
best_tree <- rtree(20)
sample_trees <- list()
for(i in 1:100){ sample_trees[[i]] <- rSPR(best_tree, 1)}
class(sample_trees) <- "multiPhylo"
best_tree <- clade_support(sample_trees, best_tree)
plot(best_tree)
nodelabels(text = best_tree$node.label)
}
