% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{collapse.thresh}
\alias{collapse.thresh}
\title{Collapse nodes below a threshold value using node labels.}
\usage{
collapse.thresh(tree, threshold)
}
\arguments{
\item{tree}{a tree object of class 'phylo'.}

\item{threshold}{a numeric value. All nodes with a numeric node.label less than the threshold will be collapsed.}
}
\value{
a tree object of class 'phylo'.
}
\description{
Function to collapse all nodes of a phylo object with numeric node labels that are less than user-specified threshold value.
}
\details{
This function requires the phylo object to have numeric node labels specified (e.g. tree$node.label).
}
\examples{
## simulate two groups of trees
set.seed(8)
t1 <- rtree(10)
trees1 <- t1
for(i in 1:49){ 
  trees1 <- c(trees1, rSPR(t1, 1))
}
t2 <- rSPR(t1, 3)
t2 <- multi2di(t2)
trees2 <- t2
for(i in 1:49){ 
  trees2 <- c(trees2, rSPR(t2, 1))
}
trees <- c(trees1, trees2)
## get maximum clade credibility tree
mcc <- maxCladeCred(trees)
plot(mcc)
## calculate clade support values
mcc <- clade.support(trees, mcc, digits = 2)
nodelabels(text = mcc$node.label)
## collapse nodes with less than 50\% support 
mcc_collapsed <- collapse.thresh(mcc, 0.5)
plot(mcc_collapsed)
nodelabels(text = mcc_collapsed$node.label)
}
