% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{density.timescale_calibration}
\alias{density.timescale_calibration}
\title{Evaluate the probability density of a node calibration}
\usage{
\method{density}{timescale_calibration}(x, age, log = FALSE, ...)
}
\arguments{
\item{x}{An object of class `"timescale_calibration"`.}

\item{age}{A numeric vector of ages (Ma).}

\item{log}{Logical. Should the log-density be returned?}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A numeric vector of probability densities.
}
\description{
Evaluates the probability density function of a node calibration at
one or more ages.
}
