% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\docType{data}
\name{dino_data}
\alias{dino_data}
\title{Dinosaur-bird integument data}
\format{
time tree of class 'phylo' representing the relationships of the 96 archosaur taxa (predominantly theropods). Integument character data ('morph') is included as an object of class 'data.frame'. Character state decriptions ('state_labels') for each of the three integument characters are included as a list.
}
\usage{
data(dino_data)
}
\description{
Phylogeny and associated integument character data modified from Cockx et al (2024, in prep).
}
\examples{
data(dino_data)
plot(dino_data, cex = 0.5, no.margin = T)
head(dino_data$morph)
}
\references{
Cockx et al (2024, in prep)
}
\keyword{datasets}
