% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{drop_otus}
\alias{drop_otus}
\title{Drop OTUs from a phylogenetic dataset}
\usage{
drop_otus(x, taxa)
}
\arguments{
\item{x}{An object of class \code{"nexdat"} or
\code{"multi_nexdat"}.}

\item{taxa}{A character vector of taxon names to remove.}
}
\value{
An object of the same class as \code{x} with the specified
taxa removed.
}
\description{
Remove one or more OTUs (taxa) from a \code{"nexdat"} or
\code{"multi_nexdat"} object.
}
\details{
This function removes OTUs from phylogenetic character matrices while
preserving partition structure and metadata.

For objects of class \code{"multi_nexdat"}, taxa are removed from all
partitions simultaneously. Partition metadata stored in the
\code{"part.info"} attribute are retained.

Taxa names must exactly match those present in the dataset.
}
\examples{
data(Lavoue2016)

x <- cat_data(
    Lavoue2016$standard,
    Lavoue2016$dna,
    use.part.info = TRUE
)

## Remove taxa
y <- drop_otus(
    x,
    taxa = c("Danio_rerio", "Esox_lucius")
)

summary(y)

}
