% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{exp_calib}
\alias{exp_calib}
\title{Generate an exponential node calibration}
\usage{
exp_calib(age_min, age_max, position = 0.95)
}
\arguments{
\item{age_min}{Hard minimum fossil calibration.}

\item{age_max}{Soft maximum fossil calibration.}

\item{position}{Cumulative probability associated with the soft
maximum age (e.g. 0.95).}
}
\value{
An object of class `"timescale_calibration"`.
}
\description{
Constructs an exponential node calibration from a hard minimum and
soft maximum age.
}
