% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{fit_timescale_ml}
\alias{fit_timescale_ml}
\title{Fit a maximum-likelihood time-scaling model}
\usage{
fit_timescale_ml(model, method = "BFGS", control = list())
}
\arguments{
\item{model}{A \code{"timescale_ml"} object.}

\item{method}{Optimisation method passed to \code{\link[stats]{optim}}.}

\item{control}{A list of control arguments passed to \code{\link[stats]{optim}}.}
}
\value{
An object of class \code{"timescale_ml_fit"}.
}
\description{
Estimates divergence times under the specified rate model.
}
