% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{format_mrbayes}
\alias{format_mrbayes}
\title{Format a node calibration for MrBayes}
\usage{
format_mrbayes(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A character string containing the corresponding MrBayes
calibration distribution.
}
\description{
Converts a node calibration to the corresponding MrBayes
calibration distribution.
}
\examples{
calib <- gamma_calib(
    age_min = 457.5,
    age_max = 636.1,
    shape = 2
)

format_mrbayes(calib)

calib <- exp_calib(
    age_min = 457.5,
    age_max = 636.1
)

format_mrbayes(calib)

calib <- uniform_calib(
    age_min = 457.5,
    age_max = 636.1
)

format_mrbayes(calib)
}
