% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{gamma_calib}
\alias{gamma_calib}
\title{Generate a gamma node calibration}
\usage{
gamma_calib(age_min, age_max, shape, position = 0.95)
}
\arguments{
\item{age_min}{Hard minimum fossil calibration.}

\item{age_max}{Soft maximum fossil calibration.}

\item{shape}{Gamma shape parameter.}

\item{position}{Cumulative probability associated with the soft
maximum age.}
}
\value{
An object of class `"timescale_calibration"`.
}
\description{
Constructs a gamma node calibration from a hard minimum and
soft maximum age.
}
\examples{
## Vertebrate calibration from Benton et al. (2015):
## hard minimum = 457.5 Ma, soft maximum = 636.1 Ma.

par(mfrow = c(3, 1), mar = c(2.2, 2.2, 1.5, 1.5))

## Good fossil record with most probability concentrated close to
## the hard minimum.
calib <- gamma_calib(
    age_min = 457.5,
    age_max = 636.1,
    shape = 1,
    position = 0.95
)
plot(calib, xlim = c(400, 1000))

## Good fossil record with probability concentrated further from
## the hard minimum.
calib <- gamma_calib(
    age_min = 457.5,
    age_max = 636.1,
    shape = 3,
    position = 0.95
)
plot(calib, xlim = c(400, 1000))

## Poor fossil record with a long tail beyond the soft maximum.
calib <- gamma_calib(
    age_min = 457.5,
    age_max = 636.1,
    shape = 3,
    position = 0.75
)
plot(calib, xlim = c(400, 1000))
}
