% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{get.contrast}
\alias{get.contrast}
\title{Get contrast matrix.}
\usage{
get.contrast(morph)
}
\arguments{
\item{morph}{a character vector, list or data.frame coded in 'standard' format (e.g. morphological data).}
}
\value{
A contrast matrix used by the phangorn function phyDat() to interpret standard categorical data.
}
\description{
Function to automatically extract a contrast matrix from a character vector, list or data.frame of standard categorical data (e.g. morphological data).
}
\details{
This function extracts a contrast matrix from categorical characters, coded in 'standard' format (e.g. '0', '1', '2', '-', '?', '0/1'). The contrast matrix is used by the phangorn phyDat() function in order to interpret symbols denoting character state ambiguity (e.g. polymorphism, missing data etc.). Data in phyDat format can be used to estimate phylogeny in R (e.g. phangorn::parsimony(), TreeSearch::MaximizeParsimony()).
}
\examples{
## Load data
install.packages("TreeSearch")
library(TreeSearch)
data("inapplicable.datasets")
morph <- inapplicable.datasets$Vinther2008
## get contrast matrix
cm <- get.contrast(morph)
## Convert data to phyDat format
pdat <- phyDat(data = morph, type = "USER", contrast = cm)
## Run parsimony analysis
trees <- MaximizeParsimony(dataset = pdat)
## Plot consensus tree and support values
con_tree <- consensus(trees, p = 0.5)
con_tree <- root.node(con_tree, 1)
con_tree$node.label[-1] <- round(as.numeric(con_tree$node.label[-1]), 2)
plot(ladderize(con_tree), cex = 0.5, no.margin = T)
nodelabels(text = con_tree$node.label)
}
