% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{get.info}
\alias{get.info}
\title{Calculate AIC, AICc and BIC}
\usage{
get.info(loglik, k, n)
}
\arguments{
\item{loglik}{numeric. The log likelihood of the model.}

\item{k}{numeric. The number of free parameters of the model.}

\item{n}{numeric. The sample size (i.e. the number of observations).}
}
\value{
A numeric vector comprising the AIC, AICc and BIC.
}
\description{
Function to calculate the AIC, AICc and BIC of a model.
}
\examples{
## simulate some binary data with the ARD model.
tree <- rtree(50)
Q <- structure(c(-0.2, 0.8, 0.2, -0.8), dim = c(2L, 2L))
true_history <- sim.history(tree, Q)
## Estimate Q using ER and ARD models.
fitER <- fitMk(tree, x = true_history$states, model = "ER")
fitARD <- fitMk(tree, x = true_history$states, model = "ARD")
## Calculate information criteria. 
get.info(fitER$logLik, 1, 50)
get.info(fitARD$logLik, 2, 50)
}
\references{
Akaike, H., 1998. Information theory and an extension of the maximum likelihood principle. In Selected papers of hirotugu akaike (pp. 199-213). New York, NY: Springer New York. \cr 
Sugiura, N., 1978. Further analysis of the data by akaike's information criterion and the finite corrections: Further analysis of the data by akaike's. Communications in Statistics-theory and Methods, 7(1), pp.13-26. \cr 
Hurvich, C.M. and Tsai, C.L., 1991. Bias of the corrected AIC criterion for underfitted regression and time series models. Biometrika, 78(3), pp.499-509.  \cr 
Schwarz, G., 1978. Estimating the dimension of a model. The annals of statistics, pp.461-464.
}
