% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{get_descendant_edges}
\alias{get_descendant_edges}
\title{Get Descendant Edges}
\usage{
get_descendant_edges(tree, current = T)
}
\arguments{
\item{tree}{an object of class 'phylo'.}

\item{current}{logical. If true, returns current edge in list of descendants.}
}
\value{
a list of length = Nedge(tree) containing the edge indices of each edge descending from edge = i.
}
\description{
Function that, for each edge in a phylo object, returns a vector of descendant edges.
}
\details{
Internal function used in anc_timeslice().
}
\examples{
t1 <- rtree(10)
get_descendant_edges(t1)

}
