% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{get_node_ages}
\alias{get_node_ages}
\title{Get Node Ages}
\usage{
get_node_ages(tree)
}
\arguments{
\item{tree}{an object of class 'phylo'.}
}
\value{
Returns a numeric vector of internal node ages, where names() are node numbers.
}
\description{
Function that estimates ancestral trait values along each edge of a phylogenetic tree, based on a continuous trait and a user-specified number of time slices.
}
\details{
This function calculates the ages of internal nodes in a phylogenetic tree with branch lengths and a specified root time. It computes the distance from the root to each internal node and converts these distances into absolute ages by subtracting them from the root time.
}
\examples{
data(vert_data)
get_node_ages(vert_data)
}
