% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\docType{data}
\name{KeatingDonoghue}
\alias{KeatingDonoghue}
\title{Character matrix of Keating & Donoghue 2016}
\format{
an object of class matrix (array).
}
\usage{
data(KeatingDonoghue)
}
\description{
Morphological taxon-by-character matrix of Keating & Donoghue 2016.
}
\examples{
data(KeatingDonoghue)
head(KeatingDonoghue)
}
\references{
Keating, J.N. and Donoghue, P.C., 2016. Histology and affinity of anaspids, and the early evolution of the vertebrate dermal skeleton. Proceedings of the Royal Society B: Biological Sciences, 283(1826), p.20152917.
}
\keyword{datasets}
