% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\docType{data}
\name{KeatingDonoghueTree}
\alias{KeatingDonoghueTree}
\title{Tree from Keating & Donoghue 2016}
\format{
an object of class "phylo".
}
\usage{
data(KeatingDonoghueTree)
}
\description{
Phylogenetic tree inferred from the morphological matrix of Keating & Donoghue 2016.
}
\examples{
data(KeatingDonoghueTree)
plot(KeatingDonoghueTree)
}
\references{
Keating, J.N. and Donoghue, P.C., 2016. Histology and affinity of anaspids, and the early evolution of the vertebrate dermal skeleton. Proceedings of the Royal Society B: Biological Sciences, 283(1826), p.20152917.
}
\keyword{datasets}
