% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\docType{data}
\name{Lavoue2016}
\alias{Lavoue2016}
\title{Total-evidence dataset from Lavoue (2016).}
\format{
A 'list' object
}
\usage{
data(Lavoue2016)
}
\description{
A list comprising morphological and molecular data partitions which make up the total-evidence phylogenetic dataset of Osteoglossiformes from Lavoue (2016).
}
\examples{
data(Lavoue2016)
combined_data <- cat_data(Lavoue2016$morph, Lavoue2016$mol, use.part.info = F) 
## Visualise 
if(!require("BiocManager", quietly = TRUE)){
	install.packages("BiocManager")
}
BiocManager::install("ComplexHeatmap")
library(ComplexHeatmap)
df <- t(as.data.frame(combined_data))
df[df == "?"] <- NA
df[df == "-"] <- NA
df[df == "n"] <- NA
states <- as.character(na.omit(unique(as.character(df))))
cols <- sample(hcl.colors(n = length(states), "zissou 1"))
names(cols) <- states
Heatmap(df[,1:1500], row_names_side = "left", col = cols, na_col = "white", name = "states")

}
\references{
Lavoué, S., 2016. Was Gondwanan breakup the cause of the intercontinental distribution of Osteoglossiformes? A time-calibrated phylogenetic test combining molecular, morphological, and paleontological evidence. Molecular Phylogenetics and Evolution, 99, pp.34-43.
}
\keyword{datasets}
