% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{make.asym.tree}
\alias{make.asym.tree}
\title{Make perfectly asymmetrical tree.}
\usage{
make.asym.tree(n = 32)
}
\arguments{
\item{n}{number of tips to generate.}
}
\value{
a tree object of class 'phylo'.
}
\description{
Function to generate a perfectly asymmetrical (unbalanced) tree with equal edge lengths and a height of 1.
}
\examples{
t <- make.asym.tree(64)
plot(t, show.tip.label = F)
}
