% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{make.sym.tree}
\alias{make.sym.tree}
\title{Make perfectly symmetrical tree.}
\usage{
make.sym.tree(n = 32)
}
\arguments{
\item{n}{number of tips to generate. Must be in the geometric sequence: a1=2, r=2 (e.g. 2, 4, 8, 16, 32...).}
}
\value{
a tree object of class 'phylo'.
}
\description{
Function to generate a perfectly symmetrical (balanced) tree with equal edge lengths and a height of 1.
}
\examples{
t <- make.sym.tree(64)
plot(t, show.tip.label = F)
}
