% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{phyDat.to.list}
\alias{phyDat.to.list}
\title{Convert phyDat to list.}
\usage{
phyDat.to.list(x)
}
\arguments{
\item{x}{an object of class 'phyDat'.}
}
\value{
a list of length t, where t is the number of taxa. Each element of the list is a character vector of length c, where c is the number of phylogenetic characters.
}
\description{
This function coverts phylogenetic data in phyDat format to list format, equivalent to the output of read.nexus.data() or read.nexdat().
}
\details{
Note that phyDat objects can be converted to matrix objects using the as.character() function.
}
\examples{
data("Laurasiatherian")
x <- phyDat.to.list(Laurasiatherian)
}
