% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{phyDat_to_nexdat}
\alias{phyDat_to_nexdat}
\title{Convert phyDat_to_nexdat.}
\usage{
phyDat_to_nexdat(x)
}
\arguments{
\item{x}{an object of class 'phyDat'.}
}
\value{
an object of class 'nexdat'.
}
\description{
This function coverts phylogenetic data in phyDat format to list format, equivalent to the output of read.nexus.data() or read.nexdat().
}
\details{
Note that phyDat objects can be converted to matrix objects using the as.character() function.
}
\examples{
install.packages("phangorn")
library(phangorn)
data("Laurasiatherian")
x <- phyDat_to_nexdat(Laurasiatherian)
}
