% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{plot.timescale_calibration}
\alias{plot.timescale_calibration}
\title{Plot a node calibration}
\usage{
\method{plot}{timescale_calibration}(x, from = x$age_min, to = NULL, n = 1000, ...)
}
\arguments{
\item{x}{An object of class `"timescale_calibration"`.}

\item{from}{Lower age limit. Defaults to the hard minimum age.}

\item{to}{Upper age limit. Defaults to the 99.9th percentile of the
calibration distribution.}

\item{n}{Number of points used to draw the curve.}

\item{...}{Further graphical arguments passed to [plot()].}
}
\value{
Invisibly returns `x`.
}
\description{
Plots the probability density of a node calibration.
}
