% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{print.timescale_calibration}
\alias{print.timescale_calibration}
\title{Print a node calibration}
\usage{
\method{print}{timescale_calibration}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class `"timescale_calibration"`.}

\item{digits}{Number of significant digits to print.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The calibration object, invisibly.
}
\description{
Prints a summary of a node calibration, including its parameters
and summary statistics.
}
