% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{quantile.timescale_calibration}
\alias{quantile.timescale_calibration}
\title{Quantiles of a node calibration}
\usage{
\method{quantile}{timescale_calibration}(x, probs = c(0, 0.25, 0.5, 0.75, 1), ...)
}
\arguments{
\item{x}{An object of class `"timescale_calibration"`.}

\item{probs}{Numeric vector of probabilities.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A numeric vector of quantiles (Ma).
}
\description{
Computes quantiles of a node calibration distribution.
}
