% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{quick_pars}
\alias{quick_pars}
\title{Quick Parsimony}
\usage{
quick_pars(data, outgroup = 1, plot = T, ...)
}
\arguments{
\item{data}{An object of class phyDat containing characters.}

\item{outgroup}{string that matches the taxon to root the tree on.}

\item{plot}{logical. If true, plots the most parsimonious tree.}

\item{...}{Further arguments to be passed to pratchet().}
}
\value{
An object of class phylo.
}
\description{
Function to conduct a quick and dirty parsimony search. Perfect for teaching.
}
\details{
This function is designed for a quick and dirty parsimony search. It will use the default parameter of the phangorn function pratchet() for conducting a tree search. Users are advised to read the help page for pratchet if they wish to conduct a more exhaustive parsimony search. Other alternatives are availible (e.g. TreeSearch).
}
\examples{
## Load data
data(KeatingDonoghue)
## convert to phyDat
cont <- get_contrast(KeatingDonoghue)
pdat <- phyDat(KeatingDonoghue, type = "USER", contrast = cont)
## Run a quick parsimony search
tree <- quick_pars(pdat, plot = T, outgroup = "Cephalochordata")

#' @export 

}
