% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{read.nexdat}
\alias{read.nexdat}
\title{Read nexus data file.}
\usage{
read.nexdat(file, use.part.info = F)
}
\arguments{
\item{file}{a file name specified by either a variable of mode character, or a double-quoted string.}

\item{use.part.info}{logical. If true, matrices of different data types will be stored separately. If true, all data types will be concatenated.}
}
\value{
a list of length t, where t is the number of taxa. Each element of the list is a character vector of length c, where c is the number of phylogenetic characters.
}
\description{
This function is a modification of the ape function read.nexus.data(). The function has been modified so that it can read MrBayes files with partitions comprising different data types (e.g. datatype=mixed).
}
\details{
The read.nexus.data() function from ape has been modified so that it can read MrBayes mixed data files, either as a single matrix with data types concatenated or as a list of matrices organised by datatype. In addition to this change, the function has been modified so that it can handle spaces in polymorphic or uncertain characters (e.g. (0 1), {0 1}). Due to this change, the function no longer checks if there are spaces in taxon names. Please ensure that spaces in taxon names are removed prior to using this function as they will cause the data to be read incorrectly.
}
\examples{
## Use read.nexdat to read a data file in NEXUS format into object x
## Not run: x <- read.nexdat("file.nex")
}
