% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{root.node}
\alias{root.node}
\title{Root a tree below a user-specified node.}
\usage{
root.node(tree, node)
}
\arguments{
\item{tree}{a tree object of class 'phylo'.}

\item{node}{number of the node or tip descending from the target branch in tree$edge. Alternatively, the user can specify a numeric vector of node/tip numbers, in which case the function will attempt to root the tree on the edge descending from the most recent common ancestor of the specified nodes/tips.}
}
\value{
A tree object of class 'phylo'.
}
\description{
Function to root a tree at the midpoint of the branch ancestral to the user-specified node.
}
\details{
This function is a wrapper for the phytools function reroot(), which is more flexible as you can specify a position on an edge at which to root the tree. However, 99 times out of 100, I want to root at the midpoint of an edge. This function does this automatically. It also allows for rooting a tree with no edge.lengths, which returns an error when attempted with reroot(). See also the castor function root_in_edge(), which is similar to reroot().
}
\examples{
## Example with edge lengths
t <- rtree(10)
plot(t, label.offset = 0.1)
tiplabels()
t2 <- root.node(t, 2)
plot(t2, label.offset = 0.1)
tiplabels()
## Example without edge lengths
t3 <- t
t3$edge.length <- NULL
plot(t3, label.offset = 0.1)
tiplabels()
t4 <- root.node(t3, 2)
plot(t4, label.offset = 0.1)
tiplabels()
}
