% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{root_multi}
\alias{root_multi}
\title{Root multiple trees below a user-specified node.}
\usage{
root_node(trees, node)
}
\arguments{
\item{trees}{a tree object of class 'multiPhylo'.}

\item{node}{number of the node or tip descending from the target branch in tree$edge. Alternatively, the user can specify a numeric vector of node/tip numbers, in which case the function will attempt to root the tree on the edge descending from the most recent common ancestor of the specified nodes/tips.}
}
\value{
a tree object of class 'phylo'.
}
\description{
Function to root all trees in a multiPhylo object at the midpoint of a branch ancestral to a user-specified node.
}
\details{
see help for root_node(). This function assumes that tip labels are identical in each tree of the multiPhylo object. Internal node labels will differ between trees of different topologies, thus it is recommended that multiPhylo objects are rooted by specifying a tip or multiple tips.
}
\examples{
trees <- rmtree(100, 10)
trees2 <- root_multi(trees, 2)
}
