% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{SB}
\alias{SB}
\title{Calculate shared bipartitions between trees.}
\usage{
SB(tree1, tree2)
}
\arguments{
\item{tree1}{an object of class 'phylo'.}

\item{tree2}{an object of class 'phylo' or 'multiPhylo'.}
}
\value{
A numeric value equal to the number of shared bipartitions between the two trees.
}
\description{
Function to calculate the number of shared bipartitions between a reference tree and a single or multiple comparsion tree(s).
}
\details{
This is a symmetric measure and is closely related to the Robinson-Foulds distance.
}
\examples{
t1 <- rtree(20)
t2 <- rSPR(t1, 4)
SB(t1, t2)
}
\references{
Robinson, D.F. and Foulds, L.R., 1981. Comparison of phylogenetic trees. Mathematical biosciences, 53(1-2), pp.131-147.
}
