% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{summary.nexdat}
\alias{summary.nexdat}
\title{Summarise a nexdat object}
\usage{
\method{summary}{nexdat}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"nexdat"}.}
}
\value{
An object of class \code{"summary_nexdat"} containing summary
statistics for the dataset and per-taxon missing data information.
}
\description{
Summarise taxon occupancy, missing data and gap proportions for a
phylogenetic character matrix.
}
\details{
This function calculates:
\itemize{
  \item the number of taxa
  \item the number of characters
  \item the proportion of missing characters per taxon
  \item the proportion of gaps per taxon
}

Missing characters are defined as:
\code{"?"}, \code{"n"} and \code{"x"}.

Gap characters are defined as:
\code{"-"}.
}
\examples{
data(Lavoue2016)

summary(Lavoue2016$dna)

}
