% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{summary.timescale_calibration}
\alias{summary.timescale_calibration}
\title{Summarise a node calibration}
\usage{
\method{summary}{timescale_calibration}(object, ...)
}
\arguments{
\item{object}{An object of class `"timescale_calibration"`.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The calibration object, invisibly.
}
\description{
Prints a summary of a node calibration, including its parameters
and summary statistics.
}
