% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{total.garbage}
\alias{total.garbage}
\title{Calculate the likelihood of the 'Total Garbage' model}
\usage{
total.garbage(tip_states)
}
\arguments{
\item{tip_states}{an object of class 'character'.}
}
\value{
A numeric value equal to the likelihood of the 'Total Garbage' model given tip_states.
}
\description{
Function to calculate the likelihood of the 'Total Garbage' model from Harmon (2019).
}
\details{
The 'Total Garbage' model from Harmon (2019) is designed to test if your data provide no informatio about historical patterns of character change. It is equivalent to drawing states at random from a hat. Consider a set of tip_states and an evolutionary model with transition matrix Q. If likeihood(Q) given tip_states is similar to likelihood(Total Garbage) given tip_states, then tip_states contains little historical information from which to infer state changes. However if likeihood(Q) given tip_states is greater than likelihood(Total Garbage) given tip_states, tip_states does contain some information about historical character changes.
}
\examples{
## simulate some data with slow evolutionary rates
tree <- rtree(50)
Q <- matrix(c(-0.2, 0.2, 0.2, -0.2), nrow = 2, ncol = 2)
true_history <- sim.history(tree, Q)
## Estimate Q.
fitMK <- fitMk(tree, x = true_history$states, model = "ER")
## Compare likelihoods. 
fitMK$logLik
total.garbage(true_history$states)
fitMK$logLik > total.garbage(true_history$states)
## simulate some data with very fast evolutionary rates
tree <- rtree(50)
Q <- matrix(c(-100, 100, 100, -100), nrow = 2, ncol = 2)
true_history <- sim.history(tree, Q)
## Estimate Q.
fitMK <- fitMk(tree, x = true_history$states, model = "ER")
## Compare likelihoods. 
fitMK$logLik
total.garbage(true_history$states)
fitMK$logLik > total.garbage(true_history$states)
}
\references{
Harmon, L., 2019. Phylogenetic comparative methods: learning from trees. https://lukejharmon.github.io/pcm/
}
