% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon-package.R
\docType{package}
\name{treesurgeon-package}
\alias{treesurgeon}
\alias{treesurgeon-package}
\title{treesurgeon: treesurgeon: Tools and utilities for phylogenetic analyses in R}
\description{
Tools and timesavers for tinkering with phylogenetic trees and data in R. TO DO! Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
