% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{u.edge.lengths}
\alias{u.edge.lengths}
\title{Get ultrametric edge lengths}
\usage{
u.edge.lengths(tree)
}
\arguments{
\item{tree}{a tree object of class 'phylo'.}
}
\value{
a tree object of class 'phylo' with ultrametric edge lengths.
}
\description{
Function to generate ultrametric edge lengths for tree plots.
}
\details{
This function is equivalent to plot.tree(node.depth = 2) except that a vector of edge.lengths are included in the output 'phylo' object.
}
\examples{
t1 <- rtree(30);
plot(t1);
t2 <- u.edge.lengths(t1);
plot(t2);
}
