% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{UB}
\alias{UB}
\title{Calculate unique bipartitions in a tree.}
\usage{
UB(tree1, tree2)
}
\arguments{
\item{tree1}{an object of class 'phylo'.}

\item{tree2}{an object of class 'phylo'.}
}
\value{
A numeric value equal to the number of  bipartitions between the two trees.
}
\description{
Function to calculate the number of bipartitions in a reference tree, that are not found in a comparsion tree.
}
\details{
This is an asymmetric measure and is closely related to the Robinson-Foulds distance. UB(tree1, tree2) + UB(tree2, tree1) = RF.dist.
}
\examples{
t1 <- rtree(20)
t2 <- rSPR(t1, 4)
UB(t1, t2)
}
\references{
Robinson, D.F. and Foulds, L.R., 1981. Comparison of phylogenetic trees. Mathematical biosciences, 53(1-2), pp.131-147.
}
