% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{uniform_calib}
\alias{uniform_calib}
\title{Generate a uniform node calibration}
\usage{
uniform_calib(age_min, age_max)
}
\arguments{
\item{age_min}{Hard minimum fossil calibration.}

\item{age_max}{Hard maximum fossil calibration.}
}
\value{
An object of class `"timescale_calibration"`.
}
\description{
Constructs a uniform node calibration with a hard minimum and
hard maximum age.
}
\examples{
## Uniform calibration between 457.5 Ma and 636.1 Ma.

calib <- uniform_calib(
    age_min = 457.5,
    age_max = 636.1
)

plot(calib, xlim = c(400, 700))

}
