% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\docType{data}
\name{vert_data}
\alias{vert_data}
\title{Early vertebrate skeleton data}
\format{
time tree of class 'phylo' representing the relationships of the 71 vertebrate taxa. Associated morphological character data ('morph') is included as an object of class 'data.frame'.
}
\usage{
data(vert_data)
}
\description{
Phylogeny and associated morphological character data modified from Keating & Donoghue (2016), O'shea et al (2019), de Buffrénil et al (2016) and Witzmann (2009).
}
\examples{
data(vert_data)
plot(vert_data, cex = 0.5, no.margin = T)
head(vert_data$morph)
}
\references{
Keating, J.N. and Donoghue, P.C., 2016. Histology and affinity of anaspids, and the early evolution of the vertebrate dermal skeleton. Proceedings of the Royal Society B: Biological Sciences, 283(1826), p.20152917. \cr 
O'Shea, J., Keating, J.N. and Donoghue, P.C., 2019. The dermal skeleton of the jawless vertebrate Tremataspis mammillata (Osteostraci, stem‐Gnathostomata). Journal of morphology, 280(7), pp.999-1025. \cr 
de Buffrénil, V., Clarac, F., Canoville, A. and Laurin, M., 2016. Comparative data on the differentiation and growth of bone ornamentation in gnathostomes (Chordata: Vertebrata). Journal of Morphology, 277(5), pp.634-670. \cr 
Witzmann, F, 2009. Comparative histology of sculptured dermal bones in basal tetrapods, and the implications for the soft tissue dermis. Palaeodiversity, 2(233), p.e270.
}
\keyword{datasets}
