% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{cat_data}
\alias{cat_data}
\title{Combine phylogenetic data partitions}
\usage{
cat_data(..., use.part.info = FALSE, part.names = NULL)
}
\arguments{
\item{...}{Two or more phylogenetic data objects. Inputs may be:
\itemize{
  \item lists of named character vectors (e.g. output from
  \code{ape::read.nexus.data()})
  \item objects of class \code{"nexdat"} or \code{"multi_nexdat"}
}}

\item{use.part.info}{Logical. If \code{TRUE}, partition metadata are stored
in the returned object as the \code{"part.info"} attribute.}

\item{part.names}{Optional character vector giving partition names. If not
supplied, partition names are inherited from the input object names.}
}
\value{
An object of class \code{"multi_nexdat"} containing one list per
partition. Partition metadata are stored in the \code{"part.info"}
attribute when \code{use.part.info = TRUE}.
}
\description{
Concatenate phylogenetic character data from multiple partitions into a
single multi-partition object. Input data can include individual partitions
(lists of named character vectors) or existing objects of class
`nexdat` or `"multi_nexdat"`. Missing taxa are automatically padded with missing data
(`"?"`) so that all partitions contain the same set of OTUs.
}
\details{
This function is designed for combining phylogenetic datasets with partially
overlapping taxon sampling, such as AA, DNA and morphological partitions. Taxa absent from a partition are automatically coded as missing.

Partition datatypes are automatically inferred where possible and currently
include:
\itemize{
  \item \code{"dna"}
  \item \code{"rna"}
  \item \code{"protein"}
  \item \code{"standard"}
}

Existing \code{"multi_nexdat"} objects can be extended by supplying them as
inputs alongside additional partitions.

Taxon names must match exactly across input datasets for sequences to be
combined correctly.
}
\examples{
#' data(Lavoue2016) 
combined_data <- cat_data(Lavoue2016$standard, Lavoue2016$dna, use.part.info = F) 
## Visualise 
if (!require("BiocManager", quietly = TRUE)) { 
 install.packages("BiocManager") 
} 
BiocManager::install("ComplexHeatmap") 
library(ComplexHeatmap) 
df <- t(as.data.frame(combined_data)) 
df <- tolower(df) 
df[df == "?"] <- NA 
df[df == "-"] <- NA
df[df == "n"] <- NA #
states <- as.character(na.omit(unique(as.character(df)))) 
cols <- rep("x", length(states)) 
names(cols) <- states 
morph_states <- suppressWarnings(which(is.na(as.numeric(names(cols))) == F)) 
mol_states <- which(names(cols) \%in\% c("a", "c", "g", "t")) 
cols[morph_states] <- hcl.colors(n = length(morph_states), palette = "Hawaii") 
cols[mol_states] <- hcl.colors(n = length(mol_states), palette = "zissou") 
cols[-c(mol_states, morph_states)] <- hcl.colors(n = length(cols[-c(mol_states, morph_states)]), palette = "Cividis")
Heatmap(df[, 1:1500], row_names_side = "left", col = cols, na_col = "white", name = "states")
}
