% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{cat_data}
\alias{cat_data}
\title{Combine phylogenetic data}
\usage{
cat_data(..., use.part.info = F)
}
\arguments{
\item{...}{two or more objects of class 'list' comprising named vectors of phylogenetic characters.}

\item{use.part.info}{logical. If TRUE, the function resturns a seperate list for each input list.}
}
\value{
A list, or multiple lists, of phylogenetic data of length T, where T equals the sum of unique taxa in the input lists.
}
\description{
Function to combine lists of phylogenetic data (e.g. those output by the ape function read.nexus.data()).
}
\details{
This function can be used to combine seperate lists of phylogenetic characters with non-overlapping taxa. If a taxon does not occur in all input lists, it will be coded as missing '?' for those input list(s). Can be used, for example, to combine molecular and morphological datasets with non-overlapping taxa. Note that taxa must be named identically in each input for the data to be combined succesfully.
}
\examples{
data(Lavoue2016)
combined_data <- cat_data(Lavoue2016$morph, Lavoue2016$mol, use.part.info = F) 
## Visualise 
if(!require("BiocManager", quietly = TRUE)){
	install.packages("BiocManager")
}
BiocManager::install("ComplexHeatmap")
library(ComplexHeatmap)
df <- t(as.data.frame(combined_data))
df[df == "?"] <- NA
df[df == "-"] <- NA
df[df == "n"] <- NA
states <- as.character(na.omit(unique(as.character(df))))
cols <- sample(hcl.colors(n = length(states), "zissou 1"))
names(cols) <- states
Heatmap(df[,1:1500], row_names_side = "left", col = cols, na_col = "white", name = "states")

}
