% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{exp_calib}
\alias{exp_calib}
\title{Generate an exponential node calibration}
\usage{
exp_calib(age_min, age_max, position = 0.95)
}
\arguments{
\item{age_min}{Hard minimum fossil calibration.}

\item{age_max}{Soft maximum fossil calibration.}

\item{position}{Cumulative probability associated with the soft
maximum age (e.g. 0.95).}
}
\value{
An object of class `"timescale_calibration"`.
}
\description{
Constructs an exponential node calibration from a hard minimum and
soft maximum age.
}
\examples{
## Vertebrate calibration from Benton et al. (2015):
## hard minimum = 457.5 Ma, soft maximum = 636.1 Ma.

par(mfrow = c(2, 1), mar = c(2.2, 2.2, 1.5, 1.5))

## Good fossil record (95\% of the probability lies below the soft maximum)
calib <- exp_calib(
    age_min = 457.5,
    age_max = 636.1,
    position = 0.95
)
print(calib)
plot(calib, xlim = c(400, 1000))

## Poor fossil record (75\% of the probability lies below the soft maximum)
calib <- exp_calib(
    age_min = 457.5,
    age_max = 636.1,
    position = 0.75
)
plot(calib, xlim = c(400, 1000))
}
