% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{exp_calib}
\alias{exp_calib}
\title{Generate an exponential node calibration}
\usage{
exp_calib(age_min, age_max, position, xlim = NULL, ylim = NULL)
}
\arguments{
\item{age_min}{hard minimum fossil calibration.}

\item{age_max}{soft maximum fossil calibration.}

\item{position}{numeric value between 0 and 1, which determines the percentile of the soft maximum age.}

\item{xlim}{numeric vector defining the dimensions of the x axis. If NULL, xlim is determined automatically.}

\item{ylim}{numeric vector defining the dimensions of the y axis. If NULL, ylim is determined automatically.}
}
\value{
a histogram of the prior distribution with labelled soft maximum and hard minimum fossil calibrations. The function also prints the mrbayes command for defining the calibration.
}
\description{
Function to generate an exponential node calibration for Bayesian molecular clock analysis.
}
\examples{
##vertebrate calibration from Benton et al (2015): hard min = 457.5, soft max = 636.1.
par(mar = c(2.2,2.2,1.5,1.5))
par(mfrow=c(2,1))
##plot calibration assuming a very good fossil record (i.e. low probability the age of the node is older than the soft maximum)
exp_calib(457.5, 636.1, 0.99, xlim = c(0, 1000), ylim = c(0, 0.025))
##plot calibration assuming a poor fossil record (i.e. high probability the age of the node is older than the soft maximum)
exp_calib(457.5, 636.1, 0.75, xlim = c(0, 1000), ylim = c(0, 0.025))
}
\references{
Benton, M.J., Donoghue, P.C., Asher, R.J., Friedman, M., Near, T.J. and Vinther, J., 2015. Constraints on the timescale of animal evolutionary history.
}
