% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{gamma_calib}
\alias{gamma_calib}
\title{Generate a gamma node calibration}
\usage{
gamma_calib(age_min, age_max, shape, position = 0.95)
}
\arguments{
\item{age_min}{Hard minimum fossil calibration.}

\item{age_max}{Soft maximum fossil calibration.}

\item{shape}{Gamma shape parameter.}

\item{position}{Cumulative probability associated with the soft
maximum age.}
}
\value{
An object of class `"timescale_calibration"`.
}
\description{
Constructs a gamma node calibration from a hard minimum and
soft maximum age.
}
