% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\docType{data}
\name{Lavoue2016}
\alias{Lavoue2016}
\title{Total-evidence dataset from Lavoue (2016).}
\format{
A 'list' object
}
\usage{
data(Lavoue2016)
}
\description{
A list comprising morphological and molecular data partitions which make up the total-evidence phylogenetic dataset of Osteoglossiformes from Lavoue (2016).
}
\examples{
data(Lavoue2016)
combined_data <- cat_data(Lavoue2016$morph, Lavoue2016$mol, use.part.info = F) 
## Visualise 
if(!require("BiocManager", quietly = TRUE)){
	install.packages("BiocManager")
}
BiocManager::install("ComplexHeatmap")
library(ComplexHeatmap)
df <- t(as.data.frame(combined_data))
df <- tolower(df)
df[df == "?"] <- NA
df[df == "-"] <- NA
df[df == "n"] <- NA
states <- as.character(na.omit(unique(as.character(df))))
cols <- rep("x", length(states))
names(cols) <- states
morph_states <- suppressWarnings(which(is.na(as.numeric(names(cols))) == F))
mol_states <- which(names(cols) \%in\% c("a", "c", "g", "t"))
cols[morph_states] <- hcl.colors(n = length(morph_states), palette = "Hawaii")
cols[mol_states] <- hcl.colors(n = length(mol_states), palette = "zissou")
cols[-c(mol_states, morph_states)] <- hcl.colors(n = length(cols[-c(mol_states, morph_states)]), palette = "Cividis")
Heatmap(df[,1:1500], row_names_side = "left", col = cols, na_col = "white", name = "states")

}
\references{
Lavoué, S., 2016. Was Gondwanan breakup the cause of the intercontinental distribution of Osteoglossiformes? A time-calibrated phylogenetic test combining molecular, morphological, and paleontological evidence. Molecular Phylogenetics and Evolution, 99, pp.34-43.
}
\keyword{datasets}
