% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{loo_cv}
\alias{loo_cv}
\title{Leave-One-Out Cross-Validation}
\usage{
loo_cv(
  tree,
  x,
  model = "ER",
  fixedQ = NULL,
  tips = seq(Ntip(tree)),
  type = "marginal",
  ...
)
}
\arguments{
\item{tree}{an object of class 'phylo'.}

\item{x}{a vector of tip values for species; names(x) should be the species names.}

\item{model}{a character string containing the model or a transition model specified in the form of a matrix. See ace for more details.}

\item{fixedQ}{fixed value of transition matrix Q, if one is desired.}

\item{tips}{Index of tips used for cross validation. By default this is set to include all tips of the tree.}

\item{type}{determines the reconstruction type. Either "joint" or "marginal".}

\item{...}{optional arguments, including pi, the prior distribution at the root node (defaults to pi="equal"). Other options for pi include pi="fitzjohn" (which implements the prior distribution of FitzJohn et al. 2009), pi="estimated" (which finds the stationary distribution of state frequencies and sets that as the prior), or an arbitrary prior distribution specified by the user.}
}
\value{
A numeric vector reporting the mean Raw error, the Brier score and the mean log likelihood of the model.
}
\description{
Function to conduct leave-one-out cross-validation on an ancestral state estimation model.
}
\details{
This function performs leave-one-out cross-validation using phytools::fitMk() and phytools::ancr(). By default, the function iterates through each tip of the tree. In each iteration, one tip state is set as uncertain, and the user-defined model is fitted to the remaining data using phytools::fitMk(). Ancestral states, including the uncertain tip state, are then estimated using phytools::ancr() with the 'tips' argument set to TRUE. The estimated uncertain tip state is compared to the true state, and both the Raw error and Brier score are recorded. Once all iterations are complete, the function returns the mean Raw error and mean Brier score for tip estimates, and the mean log likelihood of the model. Optionally, the function can be applied to a subset of tips rather than the full dataset using the tips argument.
}
\examples{
## Load data
head(vert_data$morph)
tp <- get_tip_priors(vert_data$morph[,2], extra_state = T)
colnames(tp[[1]]) <- c("bone absent", "secondary osteons absent", "secondary osteons present")
## Register parallel backend for Windows with doParallel
cl <- parallel::makeCluster(2)
doParallel::registerDoParallel(cl)
## Register parallel backend for Linux or MacOS with doParallel
nc <- parallel::detectCores()
doParallel::registerDoParallel(cores=nc)
loo_cv(tree = vert_data, x = tp[[1]], model = "ER")
loo_cv(tree = vert_data, x = tp[[1]], model = "ARD")
}
