% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{remove_part_info}
\alias{remove_part_info}
\title{remove partition information}
\usage{
remove_part_info(x)
}
\arguments{
\item{x}{a list object of subclass 'multi_nexdat'.}
}
\value{
A list object of subclass 'nexdat'.
}
\description{
Function to remove partition information from a 'multi_nexdat' object.
}
\details{
This function removes partition information from a 'multi_nexdat' object, thereby converting it into a 'nexdat' object. Sequences from seperate partitions are conctenated.
}
\examples{
data(Lavoue2016)
partitioned_data <- cat_data(Lavoue2016$standard, Lavoue2016$dna, use.part.info = T)
concatenated_data <- remove_part_info(partitioned_data)  

}
