% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{summary.multi_nexdat}
\alias{summary.multi_nexdat}
\title{summary.multi_nexdat}
\usage{
\method{summary}{multi_nexdat}(x)
}
\arguments{
\item{x}{an object of class 'multi_nexdat'.}
}
\value{
An object of class 'summary.nexdat'.
}
\description{
Function to sumarise an object of class 'multi_nexdat'.
}
\details{
This function returns the proportion of missing characters and proportion of gaps per taxon. If partitions are specidied, the function also returns the proportion of missing data, gaps, ambiguous codings (e.g. polymorphic characters, characters that are coded as uncertain between two or more states) and unamibiguous characters per taxon per partition.
}
\examples{
data(Lavoue2016)
summary(Lavoue2016)
}
