% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{summary.nexdat}
\alias{summary.nexdat}
\title{summary.nexdat}
\usage{
\method{summary}{nexdat}(x)
}
\arguments{
\item{x}{an object of class 'nexdat'.}
}
\value{
An object of class 'summary.nexdat'.
}
\description{
Function to sumarise an object of class 'nexdat'.
}
\details{
This function returns the proportion of missing characters and proportion of gaps per taxon.
}
\examples{
data(Lavoue2016) 
summary(Lavoue2016$dna)
}
