% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesurgeon.r
\name{timescale_ml}
\alias{timescale_ml}
\title{Construct a maximum-likelihood time-scaling model}
\usage{
timescale_ml(tree, tip_ages, calibrations, replace_zeros = TRUE)
}
\arguments{
\item{tree}{a phylogenetic tree of class \code{"phylo"} with branch lengths proportional to evolutionary change.}

\item{tip_ages}{a numeric vector giving the ages of the tree tips. The vector should have length \code{Ntip(tree)} and be ordered to match \code{tree$tip.label}.}

\item{calibrations}{a list of node age calibration distributions indexed by node number. Uncalibrated nodes should be assigned \code{NULL}.}

\item{replace_zeros}{logical indicating whether zero-length branches should be replaced with a small positive value before fitting.}
}
\value{
An object of class \code{"timescale_ml"} containing the phylogeny, tip ages, calibrations and information required for maximum-likelihood optimisation.
}
\description{
Constructs a model object for maximum-likelihood time-scaling of a phylogenetic tree using fossil tip ages and node age calibrations. The returned model can be passed to \code{fit_timescale_ml()} to estimate divergence times and branch-specific evolutionary rates.
}
\examples{
tree <- KeatingDonoghueTree

calibrations <- vector("list", Ntip(tree) + Nnode(tree))

calibrations[[22]] <- gamma_calib(
    age_min = 516.1,
    age_max = 636.1,
    shape = 1.2,
    position = 0.99
)

calibrations[[24]] <- gamma_calib(
    age_min = 461.1,
    age_max = 636.1,
    shape = 1,
    position = 0.99
)

model <- timescale_ml(
    tree = tree,
    tip_ages = tree$tip_ages,
    calibrations = calibrations,
    replace_zeros = TRUE
)

fit <- fit_timescale_ml(model)
plot(fit$tree)
axisPhylo()

}
